refactor: :recycle: change javascript to typescript
