+++ src/services/user.ts
@@ -1,0 +1,103 @@
+import { APIResponse, ErrorTypes, ResponseOn } from '../config/utils/response';
+import { UserEntity } from '../entities/user';
+import { UserRepository } from '../repositories/user';
+
+const response = new ResponseOn();
+const userRepository = new UserRepository();
+
+export class UserService {
+    getAll = async (): Promise<APIResponse<UserEntity[], ErrorTypes>> => {
+        try {
+            const users = await userRepository.getAll();
+
+            if (users.length === 0 || !users) {
+                return response.error('Nenhum usuário encontrado', 404);
+            }
+
+            return response.success(users, 200);
+
+        } catch (error) {
+            return response.error(error, 500);
+        }
+    };
+
+    getById = async (id: number): Promise<APIResponse<UserEntity | null, ErrorTypes>> => {
+        try {
+            if (!id) {
+                return response.error('O id é obrigatório', 400);
+            }
+
+            const user = await userRepository.getById(id);
+
+            if (!user) {
+                return response.error(`Usuário de id ${id} não encontrado`, 404);
+            }
+
+            return response.success(user, 200);
+        } catch (error) {
+            return response.error(error, 500);
+        }
+    };
+
+    create = async (name: string, email: string, password: string): Promise<APIResponse<string | null, ErrorTypes>> => {
+        try {
+            if (!name || !email || !password) {
+                return response.error('O nome, o email e a senha são obrigatórios', 400);
+            }
+
+            const checkUserExist = await userRepository.getByEmail(email);
+
+            if (!checkUserExist) {
+                return response.error('O usuário já existe', 400);
+            }
+
+            const roleIdDefault = 2;
+
+            await userRepository.create({ name, email, password, role_id: roleIdDefault });
+
+            return response.success('Usuário foi criado com sucesso', 201);
+        } catch (error) {
+            return response.error(error, 500);
+        }
+    };
+
+    update = async (id: number, name: string, email: string): Promise<APIResponse<string | null, ErrorTypes>> => {
+        try {
+            if (!id) {
+                return response.error('O id do usuário é obrigatório', 400);
+            }
+
+            const checkUserExist = await userRepository.getById(id);
+
+            if (!checkUserExist) {
+                return response.error(`Usuário de id ${id} não encontrado`, 404);
+            }
+        
+            await userRepository.update(id, { name, email });
+
+            return response.success('Usuário foi atualizado com sucesso', 200);
+        } catch (error) {
+            return response.error(error, 500);
+        }
+    };
+
+    exclude = async (id: number): Promise<APIResponse<string | null, ErrorTypes>> => {
+        try {
+            if (!id) {
+                return response.error('O id é obrigatório', 400);
+            }
+
+            const checkUserExist = await userRepository.getById(id);
+
+            if (!checkUserExist) {
+                response.error(`Usuário de id ${id} não encontrado`, 404);
+            }
+
+            await userRepository.exclude(id);
+
+            return response.success('Usuário foi excluído com sucesso', 200);
+        } catch (error) {
+            return response.error(error, 500);
+        }
+    };
+}
